Allow more fine grained control over QFont debug output
With verbosity level 0, only the resolved properties are included
in the debug output, which is useful when debugging font resolving.

With verbosity level 1, the family, size, and weights are included,
plus any resolved property that is different from the default value.

Verbosity level 2, the QDebug default, has been kept unchanged,
except removing quotes and spaces by adding nospace() and qPrintable.

Verbosity level 3 includes all properties, regardless of whether or
not they have been resolved.

Levels 1 and 3 also include the resolve mask, to aid debugging.

This gives the following results for a QFont set to 100pt:

  QFont(100pt)
  QFont(".SF NS Text", 100pt, Weight::Normal, resolveMask=SizeResolved)
  QFont(.SF NS Text,100,-1,5,50,0,0,0,0,0)
  QFont(".SF NS Text", 100pt, StyleHint::AnyStyle, StyleStrategy::PreferDefault,
    Weight::Normal, StyleNormal, underline=false, overline=false, strikeOut=false,
    fixedPitch=false, AnyStretch, kerning=true, MixedCase, letterSpacing=0
    (PercentageSpacing), PreferDefaultHinting, styleName="", resolveMask=SizeResolved)

Change-Id: Ib4aebd7346ef4a2946cc4450c12730cf7844c3bc
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>