QFormLayout: honor vertical expanding size policy
QFormLayout did not honor the vertical expanding state for widgets.
Therefore e.g. a QTextEdit was not expanded vertically when it was
inside a QFormLayout but worked fine inside a QGridLayout.
It was honored when a stretch factor was set though.
Fix it by not adding a dummy stretch item when one item is expanding.

[ChangeLog][QtWidgets][QFormLayout] Honor the vertical expanding state
of a widget inside a QFormLayout.

Fixes: QTBUG-72676
Change-Id: If4456145918afa5a10435063770cc93bb9315fbe
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>