Reorder the elements of the Android manifest to fix a lint warning
Incorrect order of elements in manifest
---------------------------------------
The <application> tag should appear after the elements which declare which
version you need, which features you need, which libraries you need, and so on.
In the past there have been subtle bugs (such as themes not getting applied
correctly) when the <application> tag appears before some of these other
elements, so it's best to order your manifest in the logical dependency order.

Change-Id: I647c6eed71b678f4df65a00fbd51135b93959e11
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>